+++ pkgs/intl4x/lib/src/datetime_format/datetime_format_4x.dart
@@ -3,7 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import '../../intl4x.dart';
-import '../locales.dart';
+import '../locale.dart';
 import '../options.dart';
 import 'datetime_format.dart';
