+++ pkgs/intl4x/lib/src/datetime_format/datetime_format_ecma.dart
@@ -7,7 +7,7 @@ import 'package:js/js_util.dart';
 
 @JS()
 import '../../intl4x.dart';
-import '../locales.dart';
+import '../locale.dart';
 import '../options.dart';
 import '../utils.dart';
 import 'datetime_format.dart';
