+++ pkgs/intl4x/lib/src/list_format/list_format_ecma.dart
@@ -6,7 +6,7 @@ import 'package:js/js.dart';
 import 'package:js/js_util.dart';
 
 import '../../intl4x.dart';
-import '../locales.dart';
+import '../locale.dart';
 import '../options.dart';
 @JS()
 import '../utils.dart';
