+++ pkgs/intl4x/lib/src/list_format/list_format_stub.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../locales.dart';
+import '../locale.dart';
 import '../options.dart';
 import 'list_format.dart';
