+++ pkgs/intl4x/lib/src/utils.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'locales.dart';
+import 'locale.dart';
 
 /// In js, locales are not written using an underscore but using a dash.
 String localeToJs(Locale locale) => locale.replaceAll('_', '-');
